lint: rename package
