+++ app/src/main/java/com/costular/atomtasks/ui/components/createtask/CreateTask.kt
@@ -1,4 +1,4 @@
-package com.costular.atomtasks.ui.components.create_task
+package com.costular.atomtasks.ui.components.createtask
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
