+++ app/src/main/java/com/costular/atomtasks/ui/components/createtask/CreateTaskExpandedState.kt
@@ -1,4 +1,4 @@
-package com.costular.atomtasks.ui.components.create_task
+package com.costular.atomtasks.ui.components.createtask
 
 import java.time.LocalDate
 import java.time.LocalTime
