+++ app/src/main/java/com/costular/atomtasks/ui/components/createtask/CreateTaskExpandedViewModel.kt
@@ -1,4 +1,4 @@
-package com.costular.atomtasks.ui.components.create_task
+package com.costular.atomtasks.ui.components.createtask
 
 import com.costular.atomtasks.ui.mvi.MviViewModel
 import java.time.LocalDate
