+++ app/src/test/java/com/costular/atomtasks/ui/components/createtask/CreateTaskExpandedViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.costular.atomtasks.ui.components.create_task
+package com.costular.atomtasks.ui.components.createtask
 
 import app.cash.turbine.test
 import com.costular.atomtasks.MviViewModelTest
